move common functions to separate package
