+++ pkg/networkpath/traceroute/tcp/utils.go
@@ -8,46 +8,14 @@ package tcp
 import (
 	"fmt"
 	"net"
-	"strconv"
+	"syscall"
 
-	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/google/gopacket"
 	"github.com/google/gopacket/layers"
 	"golang.org/x/net/ipv4"
 )
 
-const (
-	// ACK is the acknowledge TCP flag
-	ACK = 1 << 4
-	// RST is the reset TCP flag
-	RST = 1 << 2
-	// SYN is the synchronization TCP flag
-	SYN = 1 << 1
-
-	// IPProtoICMP is the ICMP protocol number
-	IPProtoICMP = 1
-	// IPProtoTCP is the TCP protocol number
-	IPProtoTCP = 6
-)
-
 type (
-	// canceledError is sent when a listener
-	// is canceled
-	canceledError string
-
-	// icmpResponse encapsulates the data from
-	// an ICMP response packet needed for matching
-	icmpResponse struct {
-		SrcIP        net.IP
-		DstIP        net.IP
-		TypeCode     layers.ICMPv4TypeCode
-		InnerSrcIP   net.IP
-		InnerDstIP   net.IP
-		InnerSrcPort uint16
-		InnerDstPort uint16
-		InnerSeqNum  uint32
-	}
-
 	// tcpResponse encapsulates the data from a
 	// TCP response needed for matching
 	tcpResponse struct {
@@ -56,25 +24,6 @@ type (
 		TCPResponse layers.TCP
 	}
 )
-
-func localAddrForHost(destIP net.IP, destPort uint16) (*net.UDPAddr, error) {
-	// this is a quick way to get the local address for connecting to the host
-	// using UDP as the network type to avoid actually creating a connection to
-	// the host, just get the OS to give us a local IP and local ephemeral port
-	conn, err := net.Dial("udp4", net.JoinHostPort(destIP.String(), strconv.Itoa(int(destPort))))
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
-	localAddr := conn.LocalAddr()
-
-	localUDPAddr, ok := localAddr.(*net.UDPAddr)
-	if !ok {
-		return nil, fmt.Errorf("invalid address type for %s: want %T, got %T", localAddr, localUDPAddr, localAddr)
-	}
-
-	return localUDPAddr, nil
-}
 
 // reserveLocalPort reserves an ephemeral TCP port
 // and returns both the listener and port because the
@@ -145,64 +94,6 @@ func createRawTCPSynBuffer(sourceIP net.IP, sourcePort uint16, destIP net.IP, de
 	return &ipHdr, packet, 20, nil
 }
 
-// parseICMP takes in an IPv4 header and payload and tries to convert to an ICMP
-// message, it returns all the fields from the packet we need to validate it's the response
-// we're looking for
-func parseICMP(header *ipv4.Header, payload []byte) (*icmpResponse, error) {
-	// in addition to parsing, it is probably not a bad idea to do some validation
-	// so we can ignore the ICMP packets we don't care about
-	icmpResponse := icmpResponse{}
-
-	if header.Protocol != IPProtoICMP || header.Version != 4 ||
-		header.Src == nil || header.Dst == nil {
-		return nil, fmt.Errorf("invalid IP header for ICMP packet: %+v", header)
-	}
-	icmpResponse.SrcIP = header.Src
-	icmpResponse.DstIP = header.Dst
-
-	var icmpv4Layer layers.ICMPv4
-	decoded := []gopacket.LayerType{}
-	icmpParser := gopacket.NewDecodingLayerParser(layers.LayerTypeICMPv4, &icmpv4Layer)
-	icmpParser.IgnoreUnsupported = true // ignore unsupported layers, we will decode them in the next step
-	if err := icmpParser.DecodeLayers(payload, &decoded); err != nil {
-		return nil, fmt.Errorf("failed to decode ICMP packet: %w", err)
-	}
-	// since we ignore unsupported layers, we need to check if we actually decoded
-	// anything
-	if len(decoded) < 1 {
-		return nil, fmt.Errorf("failed to decode ICMP packet, no layers decoded")
-	}
-	icmpResponse.TypeCode = icmpv4Layer.TypeCode
-
-	var icmpPayload []byte
-	if len(icmpv4Layer.Payload) < 40 {
-		log.Tracef("Payload length %d is less than 40, extending...\n", len(icmpv4Layer.Payload))
-		icmpPayload = make([]byte, 40)
-		copy(icmpPayload, icmpv4Layer.Payload)
-		// we have to set this in order for the TCP
-		// parser to work
-		icmpPayload[32] = 5 << 4 // set data offset
-	} else {
-		icmpPayload = icmpv4Layer.Payload
-	}
-
-	// a separate parser is needed to decode the inner IP and TCP headers because
-	// gopacket doesn't support this type of nesting in a single decoder
-	var innerIPLayer layers.IPv4
-	var innerTCPLayer layers.TCP
-	innerIPParser := gopacket.NewDecodingLayerParser(layers.LayerTypeIPv4, &innerIPLayer, &innerTCPLayer)
-	if err := innerIPParser.DecodeLayers(icmpPayload, &decoded); err != nil {
-		return nil, fmt.Errorf("failed to decode inner ICMP payload: %w", err)
-	}
-	icmpResponse.InnerSrcIP = innerIPLayer.SrcIP
-	icmpResponse.InnerDstIP = innerIPLayer.DstIP
-	icmpResponse.InnerSrcPort = uint16(innerTCPLayer.SrcPort)
-	icmpResponse.InnerDstPort = uint16(innerTCPLayer.DstPort)
-	icmpResponse.InnerSeqNum = innerTCPLayer.Seq
-
-	return &icmpResponse, nil
-}
-
 type tcpParser struct {
 	layer               layers.TCP
 	decoded             []gopacket.LayerType
@@ -216,7 +107,7 @@ func newTCPParser() *tcpParser {
 }
 
 func (tp *tcpParser) parseTCP(header *ipv4.Header, payload []byte) (*tcpResponse, error) {
-	if header.Protocol != IPProtoTCP || header.Version != 4 ||
+	if header.Protocol != syscall.IPPROTO_TCP || header.Version != 4 ||
 		header.Src == nil || header.Dst == nil {
 		return nil, fmt.Errorf("invalid IP header for TCP packet: %+v", header)
 	}
@@ -236,14 +127,6 @@ func (tp *tcpParser) parseTCP(header *ipv4.Header, payload []byte) (*tcpResponse
 	return resp, nil
 }
 
-func icmpMatch(localIP net.IP, localPort uint16, remoteIP net.IP, remotePort uint16, seqNum uint32, response *icmpResponse) bool {
-	return localIP.Equal(response.InnerSrcIP) &&
-		remoteIP.Equal(response.InnerDstIP) &&
-		localPort == response.InnerSrcPort &&
-		remotePort == response.InnerDstPort &&
-		seqNum == response.InnerSeqNum
-}
-
 func tcpMatch(localIP net.IP, localPort uint16, remoteIP net.IP, remotePort uint16, seqNum uint32, response *tcpResponse) bool {
 	flagsCheck := (response.TCPResponse.SYN && response.TCPResponse.ACK) || response.TCPResponse.RST
 	sourcePort := uint16(response.TCPResponse.SrcPort)
@@ -255,8 +138,4 @@ func tcpMatch(localIP net.IP, localPort uint16, remoteIP net.IP, remotePort uint
 		localPort == destPort &&
 		seqNum == response.TCPResponse.Ack-1 &&
 		flagsCheck
-}
-
-func (c canceledError) Error() string {
-	return string(c)
 }
