+++ pkg/networkpath/traceroute/tcp/utils_test.go
@@ -10,15 +10,14 @@ package tcp
 import (
 	"fmt"
 	"net"
-	"reflect"
 	"runtime"
 	"testing"
 
-	"github.com/google/gopacket"
-	"github.com/google/gopacket/layers"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/net/ipv4"
+
+	"github.com/DataDog/datadog-agent/pkg/networkpath/traceroute/testutils"
 )
 
 var (
@@ -113,102 +112,12 @@ func Test_createRawTCPSynBuffer(t *testing.T) {
 	assert.Equal(t, expectedPktBytes, pktBytes)
 }
 
-func Test_parseICMP(t *testing.T) {
-	ipv4Header := createMockIPv4Header(srcIP, dstIP, 1)
-	icmpLayer := createMockICMPLayer(layers.ICMPv4CodeTTLExceeded)
-	innerIPv4Layer := createMockIPv4Layer(innerSrcIP, innerDstIP, layers.IPProtocolTCP)
-	innerTCPLayer := createMockTCPLayer(12345, 443, 28394, 12737, true, true, true)
-
-	tt := []struct {
-		description string
-		inHeader    *ipv4.Header
-		inPayload   []byte
-		expected    *icmpResponse
-		errMsg      string
-	}{
-		{
-			description: "empty IPv4 layer should return an error",
-			inHeader:    &ipv4.Header{},
-			inPayload:   []byte{},
-			expected:    nil,
-			errMsg:      "invalid IP header for ICMP packet",
-		},
-		{
-			description: "missing ICMP layer should return an error",
-			inHeader:    ipv4Header,
-			inPayload:   []byte{},
-			expected:    nil,
-			errMsg:      "failed to decode ICMP packet",
-		},
-		{
-			description: "missing inner layers should return an error",
-			inHeader:    ipv4Header,
-			inPayload:   createMockICMPPacket(nil, icmpLayer, nil, nil, false),
-			expected:    nil,
-			errMsg:      "failed to decode inner ICMP payload",
-		},
-		{
-			description: "ICMP packet with partial TCP header should create icmpResponse",
-			inHeader:    ipv4Header,
-			inPayload:   createMockICMPPacket(nil, icmpLayer, innerIPv4Layer, innerTCPLayer, true),
-			expected: &icmpResponse{
-				SrcIP:        srcIP,
-				DstIP:        dstIP,
-				InnerSrcIP:   innerSrcIP,
-				InnerDstIP:   innerDstIP,
-				InnerSrcPort: 12345,
-				InnerDstPort: 443,
-				InnerSeqNum:  28394,
-			},
-			errMsg: "",
-		},
-		{
-			description: "full ICMP packet should create icmpResponse",
-			inHeader:    ipv4Header,
-			inPayload:   createMockICMPPacket(nil, icmpLayer, innerIPv4Layer, innerTCPLayer, true),
-			expected: &icmpResponse{
-				SrcIP:        srcIP,
-				DstIP:        dstIP,
-				InnerSrcIP:   innerSrcIP,
-				InnerDstIP:   innerDstIP,
-				InnerSrcPort: 12345,
-				InnerDstPort: 443,
-				InnerSeqNum:  28394,
-			},
-			errMsg: "",
-		},
-	}
-
-	for _, test := range tt {
-		t.Run(test.description, func(t *testing.T) {
-			actual, err := parseICMP(test.inHeader, test.inPayload)
-			if test.errMsg != "" {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), test.errMsg)
-				assert.Nil(t, actual)
-				return
-			}
-			require.Nil(t, err)
-			require.NotNil(t, actual)
-			// assert.Equal doesn't handle net.IP well
-			assert.Equal(t, structFieldCount(test.expected), structFieldCount(actual))
-			assert.Truef(t, test.expected.SrcIP.Equal(actual.SrcIP), "mismatch source IPs: expected %s, got %s", test.expected.SrcIP.String(), actual.SrcIP.String())
-			assert.Truef(t, test.expected.DstIP.Equal(actual.DstIP), "mismatch dest IPs: expected %s, got %s", test.expected.DstIP.String(), actual.DstIP.String())
-			assert.Truef(t, test.expected.InnerSrcIP.Equal(actual.InnerSrcIP), "mismatch inner source IPs: expected %s, got %s", test.expected.InnerSrcIP.String(), actual.InnerSrcIP.String())
-			assert.Truef(t, test.expected.InnerDstIP.Equal(actual.InnerDstIP), "mismatch inner dest IPs: expected %s, got %s", test.expected.InnerDstIP.String(), actual.InnerDstIP.String())
-			assert.Equal(t, test.expected.InnerSrcPort, actual.InnerSrcPort)
-			assert.Equal(t, test.expected.InnerDstPort, actual.InnerDstPort)
-			assert.Equal(t, test.expected.InnerSeqNum, actual.InnerSeqNum)
-		})
-	}
-}
-
 func Test_parseTCP(t *testing.T) {
-	ipv4Header := createMockIPv4Header(srcIP, dstIP, 6) // 6 is TCP
-	tcpLayer := createMockTCPLayer(12345, 443, 28394, 12737, true, true, true)
+	ipv4Header := testutils.CreateMockIPv4Header(srcIP, dstIP, 6) // 6 is TCP
+	tcpLayer := testutils.CreateMockTCPLayer(12345, 443, 28394, 12737, true, true, true)
 
 	// full packet
-	encodedTCPLayer, fullTCPPacket := createMockTCPPacket(ipv4Header, tcpLayer, false)
+	encodedTCPLayer, fullTCPPacket := testutils.CreateMockTCPPacket(ipv4Header, tcpLayer, false)
 
 	tt := []struct {
 		description string
@@ -257,7 +166,7 @@ func Test_parseTCP(t *testing.T) {
 			require.Nil(t, err)
 			require.NotNil(t, actual)
 			// assert.Equal doesn't handle net.IP well
-			assert.Equal(t, structFieldCount(test.expected), structFieldCount(actual))
+			assert.Equal(t, testutils.StructFieldCount(test.expected), testutils.StructFieldCount(actual))
 			assert.Truef(t, test.expected.SrcIP.Equal(actual.SrcIP), "mismatch source IPs: expected %s, got %s", test.expected.SrcIP.String(), actual.SrcIP.String())
 			assert.Truef(t, test.expected.DstIP.Equal(actual.DstIP), "mismatch dest IPs: expected %s, got %s", test.expected.DstIP.String(), actual.DstIP.String())
 			assert.Equal(t, test.expected.TCPResponse, actual.TCPResponse)
@@ -266,11 +175,11 @@ func Test_parseTCP(t *testing.T) {
 }
 
 func BenchmarkParseTCP(b *testing.B) {
-	ipv4Header := createMockIPv4Header(srcIP, dstIP, 6) // 6 is TCP
-	tcpLayer := createMockTCPLayer(12345, 443, 28394, 12737, true, true, true)
+	ipv4Header := testutils.CreateMockIPv4Header(srcIP, dstIP, 6) // 6 is TCP
+	tcpLayer := testutils.CreateMockTCPLayer(12345, 443, 28394, 12737, true, true, true)
 
 	// full packet
-	_, fullTCPPacket := createMockTCPPacket(ipv4Header, tcpLayer, false)
+	_, fullTCPPacket := testutils.CreateMockTCPPacket(ipv4Header, tcpLayer, false)
 
 	tp := newTCPParser()
 
@@ -281,130 +190,4 @@ func BenchmarkParseTCP(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func createMockIPv4Header(srcIP, dstIP net.IP, protocol int) *ipv4.Header {
-	return &ipv4.Header{
-		Version:  4,
-		Src:      srcIP,
-		Dst:      dstIP,
-		Protocol: protocol,
-		TTL:      64,
-		Len:      8,
-	}
-}
-
-func createMockICMPPacket(ipLayer *layers.IPv4, icmpLayer *layers.ICMPv4, innerIP *layers.IPv4, innerTCP *layers.TCP, partialTCPHeader bool) []byte {
-	innerBuf := gopacket.NewSerializeBuffer()
-	opts := gopacket.SerializeOptions{FixLengths: true, ComputeChecksums: true}
-
-	innerLayers := make([]gopacket.SerializableLayer, 0, 2)
-	if innerIP != nil {
-		innerLayers = append(innerLayers, innerIP)
-	}
-	if innerTCP != nil {
-		innerLayers = append(innerLayers, innerTCP)
-		if innerIP != nil {
-			innerTCP.SetNetworkLayerForChecksum(innerIP)
-		}
-	}
-
-	gopacket.SerializeLayers(innerBuf, opts,
-		innerLayers...,
-	)
-	payload := innerBuf.Bytes()
-
-	// if partialTCP is set, truncate
-	// the payload to include only the
-	// first 8 bytes of the TCP header
-	if partialTCPHeader {
-		payload = payload[:32]
-	}
-
-	buf := gopacket.NewSerializeBuffer()
-	gopacket.SerializeLayers(buf, opts,
-		icmpLayer,
-		gopacket.Payload(payload),
-	)
-
-	icmpBytes := buf.Bytes()
-	if ipLayer == nil {
-		return icmpBytes
-	}
-
-	buf = gopacket.NewSerializeBuffer()
-	gopacket.SerializeLayers(buf, opts,
-		ipLayer,
-		gopacket.Payload(icmpBytes),
-	)
-
-	return buf.Bytes()
-}
-
-func createMockTCPPacket(ipHeader *ipv4.Header, tcpLayer *layers.TCP, includeHeader bool) (*layers.TCP, []byte) {
-	ipLayer := &layers.IPv4{
-		Version:  4,
-		SrcIP:    ipHeader.Src,
-		DstIP:    ipHeader.Dst,
-		Protocol: layers.IPProtocol(ipHeader.Protocol),
-		TTL:      64,
-		Length:   8,
-	}
-	tcpLayer.SetNetworkLayerForChecksum(ipLayer)
-	buf := gopacket.NewSerializeBuffer()
-	opts := gopacket.SerializeOptions{FixLengths: true, ComputeChecksums: true}
-	if includeHeader {
-		gopacket.SerializeLayers(buf, opts,
-			ipLayer,
-			tcpLayer,
-		)
-	} else {
-		gopacket.SerializeLayers(buf, opts,
-			tcpLayer,
-		)
-	}
-
-	pkt := gopacket.NewPacket(buf.Bytes(), layers.LayerTypeTCP, gopacket.Default)
-
-	// return encoded TCP layer here
-	return pkt.Layer(layers.LayerTypeTCP).(*layers.TCP), buf.Bytes()
-}
-
-func createMockIPv4Layer(srcIP, dstIP net.IP, protocol layers.IPProtocol) *layers.IPv4 {
-	return &layers.IPv4{
-		SrcIP:    srcIP,
-		DstIP:    dstIP,
-		Version:  4,
-		Protocol: protocol,
-	}
-}
-
-func createMockICMPLayer(typeCode layers.ICMPv4TypeCode) *layers.ICMPv4 {
-	return &layers.ICMPv4{
-		TypeCode: typeCode,
-	}
-}
-
-func createMockTCPLayer(srcPort uint16, dstPort uint16, seqNum uint32, ackNum uint32, syn bool, ack bool, rst bool) *layers.TCP {
-	return &layers.TCP{
-		SrcPort: layers.TCPPort(srcPort),
-		DstPort: layers.TCPPort(dstPort),
-		Seq:     seqNum,
-		Ack:     ackNum,
-		SYN:     syn,
-		ACK:     ack,
-		RST:     rst,
-	}
-}
-
-func structFieldCount(v interface{}) int {
-	val := reflect.ValueOf(v)
-	if val.Kind() == reflect.Ptr {
-		val = val.Elem()
-	}
-	if val.Kind() != reflect.Struct {
-		return -1
-	}
-
-	return val.NumField()
 }
